+++ hnote/components/organisms/MainLists.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
-import { View, Text, StyleSheet } from 'react-native';
+import { View, Text, StyleSheet, FlatList } from 'react-native';
 import MenuListItem from '../molecules/MenuListItem';
+import HR from '../atoms/Hr';
 
 const MainLists = props => {
     return (
@@ -11,15 +12,24 @@ const MainLists = props => {
                 text="My Day"
             />
             <MenuListItem
-                imagePath={require('../../assets/images/sun.png')}
+                imagePath={require('../../assets/images/star.png')}
                 onPress={() => console.log('work')}
-                text="My Day"
+                text="Important"
             />
             <MenuListItem
-                imagePath={require('../../assets/images/sun.png')}
+                imagePath={require('../../assets/images/homework.png')}
                 onPress={() => console.log('work')}
-                text="My Day"
+                text="My tasks"
             />
+            <HR />
+
+            <View style={styles.newItemWrapper}>
+                <MenuListItem
+                    imagePath={require('../../assets/images/plus.png')}
+                    onPress={() => console.log('add')}
+                    text="NEW LIST"
+                />
+            </View>
         </View>
     );
 };
@@ -29,7 +39,14 @@ const styles = StyleSheet.create({
         width: '100%',
         height: '100%',
         backgroundColor: '#eee',
-        paddingLeft: 15,
+        paddingHorizontal: 15,
+    },
+    itemsWrapper: {
+        flex: 1,
+        justifyContent: 'space-between',
+    },
+    newItemWrapper: {
+        alignItems: 'center',
     },
 });